+++ src/booking.js
@@ -491,7 +491,7 @@ class RecrasBooking {
             if (isNaN(aantal)) {
                 aantal = 0;
             }
-            total += (aantal * product.verkoop);
+            total += (aantal * parseFloat(product.verkoop));
         });
         return total;
     }