+++ changelog.md
@@ -3,6 +3,7 @@
 ## master
 * Show "amount is more than allowed for online booking" message directly underneath the corresponding input instead of at the end of the form.
 * Add `recras-input-invalid` class to invalid amount inputs
+* Use minimum amount instead of number of people for "minimum amount" calculation
 
 ## 0.10.3 (2018-11-20)
 * Fix standalone contact forms