+++ changelog.md
@@ -7,6 +7,7 @@
 * Fix book button being enabled despite "I agree" checkbox not being checked
 * Show line price based on amount selected
 * Fix form validation for standalone contact forms
+* Fix subtotal when logged in
 
 ## 0.10.3 (2018-11-20)
 * Fix standalone contact forms